Fix NaNs in inverse solution of case 45 0 -45 179.572719
Addresses https://github.com/geographiclib/geographiclib-js/issues/3

JS and Java share the same directed rounding behavior for Round, namely
1.5 -> 2 while -1.5 -> -1.  This means that sincosd(+/-45) did not
behave symmetrically.

This applies the changes in commit 9416cf72 of geographiclib-js (but not
the use of x+t to determine the sign of sin(0) which was subsequently
changed back to x).

Test GeodSolve99 treats this case.